svc, config
